+++ src/eca/models.clj
@@ -36,12 +36,14 @@
     "anthropic/claude-sonnet-4.5"
     "anthropic/claude-opus-4.1"
     "anthropic/claude-opus-4.5"
+    "anthropic/claude-opus-4.6"
     "anthropic/claude-haiku-4.5"
     "anthropic/claude-sonnet-4-5-20250929"
     "anthropic/claude-sonnet-4-20250514"
     "anthropic/claude-opus-4-20250514"
     "anthropic/claude-opus-4-1-20250805"
     "anthropic/claude-opus-4-5-20251101"
+    "anthropic/claude-opus-4-6"
     "anthropic/claude-haiku-4-5-20251001"})
 
 (defn ^:private all